Refine citation finder widget and fix build
